test(oxfmt): Add tests for #22780 (#22815)
Refs: #22780
